restructure data collection scripts
